\name{poky_1964_2024}
\alias{poky_1964_2024}
\docType{data}
\title{
A Temperature Time Series for Pocatello Idaho
}
\description{
A sixty year temperature time series (1964-2024) for Pocatello, Idaho. Metadata concerning the Pocatello weather station is given in the documentation for the package \code{\link{climate}} datasets.
}
\usage{data("poky_1964_2024")}
\format{
  A data frame with 732 observations on the following 3 variables.
  \describe{
    \item{\code{Temp_C}}{Temperature in degrees C}
    \item{\code{Year}}{Year}
    \item{\code{Mos}}{Month}
  }
}
\source{
https://www.weather.gov/
}
\keyword{datasets}
